Fix : like 수정

+++ src/main/java/com/example/zzan/like/service/LikeService.java
@@ -27,7 +27,7 @@ public class LikeService {
     private final UserRepository userRepository;
 
     @Transactional
-    public ResponseDto updateAlcohol(Long targetId, boolean like) {
+    public ResponseDto increaseAlcohol(Long targetId, boolean like) {
         UserDetailsImpl currentUser = getCurrentUser();
         Long userId = currentUser.getUser().getId();
         Optional<User> targetOptional = userRepository.findById(targetId);
@@ -41,23 +41,81 @@ public class LikeService {
 
             User user = userRepository.getOne(userId);
 
-            Like existingLike = likeRepository.findByUserAndTargetUserAndLikeEnum(user, targetUser, like ? LikeEnum.LIKE : LikeEnum.DISLIKE);
+            Like existingLike = likeRepository.findByUserAndTargetUserAndLikeEnum(user, targetUser, LikeEnum.LIKE);
 
             if (existingLike == null) {
-                targetUser.setAlcohol(targetUser.getAlcohol() + (like ? 1 : -1));
+                targetUser.setAlcohol(targetUser.getAlcohol() + 1);
                 userRepository.save(targetUser);
-                Like likeEntity = new Like(user, targetUser, like ? LikeEnum.LIKE : LikeEnum.DISLIKE);
+                Like likeEntity = new Like(user, targetUser, LikeEnum.LIKE);
                 likeRepository.save(likeEntity);
                 targetUser.setAlcoholUp(like);
                 targetUser.setAlcoholDown(!like);
-                return ResponseDto.setSuccess(like ? "도수를 올렸습니다." : "도수를 내렸습니다.");
+                return ResponseDto.setSuccess("도수를 올렸습니다.");
             } else {
-                targetUser.setAlcohol(targetUser.getAlcohol() - (like ? 1 : -1));
+                if (like && existingLike.getLikeEnum() == LikeEnum.LIKE) {
+                    targetUser.setAlcohol(targetUser.getAlcohol() - 1);
+                    userRepository.save(targetUser);
+                    likeRepository.delete(existingLike);
+                    targetUser.setAlcoholUp(false);
+                    targetUser.setAlcoholDown(false);
+                    return ResponseDto.setSuccess("도수 올리기를 취소하였습니다.");
+                } else {
+                    targetUser.setAlcohol(targetUser.getAlcohol() + 2);
+                    likeRepository.delete(existingLike);
+                    Like likeEntity = new Like(user, targetUser, LikeEnum.LIKE);
+                    likeRepository.save(likeEntity);
+                    targetUser.setAlcoholUp(true);
+                    targetUser.setAlcoholDown(false);
+                    return ResponseDto.setSuccess("도수 내리기 취소 후 도수 올리기 적용");
+                }
+            }
+        } else {
+            throw new ApiException(TARGET_USER_NOT_FOUND);
+        }
+    }
+
+    @Transactional
+    public ResponseDto decreaseAlcohol(Long targetId, boolean like) {
+        UserDetailsImpl currentUser = getCurrentUser();
+        Long userId = currentUser.getUser().getId();
+        Optional<User> targetOptional = userRepository.findById(targetId);
+
+        if (targetOptional.isPresent()) {
+            User targetUser = targetOptional.get();
+
+            if (userId.equals(targetId)) {
+                throw new ApiException(NOT_ALLOWED_SELF_LIKE);
+            }
+
+            User user = userRepository.getOne(userId);
+
+            Like existingLike = likeRepository.findByUserAndTargetUserAndLikeEnum(user, targetUser, LikeEnum.DISLIKE);
+
+            if (existingLike == null) {
+                targetUser.setAlcohol(targetUser.getAlcohol() -1);
                 userRepository.save(targetUser);
-                likeRepository.delete(existingLike);
-                targetUser.setAlcoholUp(false);
-                targetUser.setAlcoholDown(false);
-                return ResponseDto.setSuccess(like ? "도수 올리기를 취소하셨습니다." : "도수 내리기를 취소하셨습니다.");
+                Like likeEntity = new Like(user, targetUser, LikeEnum.DISLIKE);
+                likeRepository.save(likeEntity);
+                targetUser.setAlcoholUp(like);
+                targetUser.setAlcoholDown(!like);
+                return ResponseDto.setSuccess("도수를 내렸습니다.");
+            } else {
+                 if (!like && existingLike.getLikeEnum() == LikeEnum.DISLIKE) {
+                    targetUser.setAlcohol(targetUser.getAlcohol() + 1);
+                    userRepository.save(targetUser);
+                    likeRepository.delete(existingLike);
+                    targetUser.setAlcoholUp(false);
+                    targetUser.setAlcoholDown(false);
+                    return ResponseDto.setSuccess("도수 내리기를 취소하였습니다.");
+                } else {
+                     targetUser.setAlcohol(targetUser.getAlcohol() - 2);
+                     likeRepository.delete(existingLike);
+                     Like likeEntity = new Like(user, targetUser, LikeEnum.DISLIKE);
+                     likeRepository.save(likeEntity);
+                     targetUser.setAlcoholUp(false);
+                     targetUser.setAlcoholDown(true);
+                     return ResponseDto.setSuccess("도수 올리기 취소 후 도수 내리기 적용");
+                 }
             }
         } else {
             throw new ApiException(TARGET_USER_NOT_FOUND);
